Updated file structure in the components folder. Updated the Links on the login page form to have routes and wrapped the app.js in a router
