+++ client/src/components/Layout/Footer/Footer.js
@@ -6,15 +6,14 @@ const Footer = () => {
             <footer className="page-footer blue lighten-1">
                 <div className="container">
                     <div className="row">
-                        <div className="col l12 m12 s12">
+                        <div className="col l12 m12 s12 center">
                             <h5 className="white-text">Footer Content</h5>
                         </div>
                     </div>
                 </div>
                 <div className="footer-copyright">
-                    <div className="container">
+                    <div className="container center">
                         © 2019 Copyright Text
-                     <a className="grey-text text-lighten-4 right" href="#!">More Links</a>
                     </div>
                 </div>
             </footer>
